requirements.txt
tweepy==4.14.0
selenium==4.15.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
psutil==7.0.0
urllib3==2.0.7
lxml==4.9.3 